finished java code coverage, and most tests for Users and Course classes

+++ backend/src/main/java/com/example/User.java
@@ -166,7 +166,7 @@ public class User {
         }
     }
 
-    public void add2Balance(int amount) {
+    public void add2Balance(double amount) {
         accountBalance += amount;
     }
